Restructure viewsets to submodules.
