+++ pydis_site/apps/api/viewsets/__init__.py
@@ -1,0 +1,17 @@
+from .bot import (  # noqa
+    BotSettingViewSet,
+    DeletedMessageViewSet,
+    DocumentationLinkViewSet,
+    InfractionViewSet,
+    NominationViewSet,
+    OffTopicChannelNameViewSet,
+    ReminderViewSet,
+    RoleViewSet,
+    SnakeFactViewSet,
+    SnakeIdiomViewSet,
+    SnakeNameViewSet,
+    SpecialSnakeViewSet,
+    TagViewSet,
+    UserViewSet
+)
+from .log_entry import LogEntryViewSet  # noqa
