+++ pydis_site/apps/api/viewsets/bot/__init__.py
@@ -1,0 +1,14 @@
+from .bot_setting import BotSettingViewSet  # noqa
+from .deleted_message import DeletedMessageViewSet  # noqa
+from .documentation_link import DocumentationLinkViewSet  # noqa
+from .infraction import InfractionViewSet  # noqa
+from .nomination import NominationViewSet  # noqa
+from .off_topic_channel_name import OffTopicChannelNameViewSet  # noqa
+from .reminder import ReminderViewSet  # noqa
+from .role import RoleViewSet  # noqa
+from .snake_fact import SnakeFactViewSet  # noqa
+from .snake_idiom import SnakeIdiomViewSet  # noqa
+from .snake_name import SnakeNameViewSet  # noqa
+from .special_snake import SpecialSnakeViewSet  # noqa
+from .tag import TagViewSet  # noqa
+from .user import UserViewSet  # noqa
