+++ pydis_site/apps/api/viewsets/bot/user.py
@@ -1,0 +1,126 @@
+from rest_framework.viewsets import ModelViewSet
+from rest_framework_bulk import BulkCreateModelMixin
+
+from pydis_site.apps.api.models.bot.user import User
+from pydis_site.apps.api.serializers import UserSerializer
+
+
+class UserViewSet(BulkCreateModelMixin, ModelViewSet):
+    """
+    View providing CRUD operations on Discord users through the bot.
+
+    ## Routes
+    ### GET /bot/users
+    Returns all users currently known.
+
+    #### Response format
+    >>> [
+    ...     {
+    ...         'id': 409107086526644234,
+    ...         'avatar': "3ba3c1acce584c20b1e96fc04bbe80eb",
+    ...         'name': "Python",
+    ...         'discriminator': 4329,
+    ...         'roles': [
+    ...             352427296948486144,
+    ...             270988689419665409,
+    ...             277546923144249364,
+    ...             458226699344019457
+    ...         ],
+    ...         'in_guild': True
+    ...     }
+    ... ]
+
+    #### Status codes
+    - 200: returned on success
+
+    ### GET /bot/users/<snowflake:int>
+    Gets a single user by ID.
+
+    #### Response format
+    >>> {
+    ...     'id': 409107086526644234,
+    ...     'avatar': "3ba3c1acce584c20b1e96fc04bbe80eb",
+    ...     'name': "Python",
+    ...     'discriminator': 4329,
+    ...     'roles': [
+    ...         352427296948486144,
+    ...         270988689419665409,
+    ...         277546923144249364,
+    ...         458226699344019457
+    ...     ],
+    ...     'in_guild': True
+    ... }
+
+    #### Status codes
+    - 200: returned on success
+    - 404: if a user with the given `snowflake` could not be found
+
+    ### POST /bot/users
+    Adds a single or multiple new users.
+    The roles attached to the user(s) must be roles known by the site.
+
+    #### Request body
+    >>> {
+    ...     'id': int,
+    ...     'avatar': str,
+    ...     'name': str,
+    ...     'discriminator': int,
+    ...     'roles': List[int],
+    ...     'in_guild': bool
+    ... }
+
+    Alternatively, request users can be POSTed as a list of above objects,
+    in which case multiple users will be created at once.
+
+    #### Status codes
+    - 201: returned on success
+    - 400: if one of the given roles does not exist, or one of the given fields is invalid
+
+    ### PUT /bot/users/<snowflake:int>
+    Update the user with the given `snowflake`.
+    All fields in the request body are required.
+
+    #### Request body
+    >>> {
+    ...     'id': int,
+    ...     'avatar': str,
+    ...     'name': str,
+    ...     'discriminator': int,
+    ...     'roles': List[int],
+    ...     'in_guild': bool
+    ... }
+
+    #### Status codes
+    - 200: returned on success
+    - 400: if the request body was invalid, see response body for details
+    - 404: if the user with the given `snowflake` could not be found
+
+    ### PATCH /bot/users/<snowflake:int>
+    Update the user with the given `snowflake`.
+    All fields in the request body are optional.
+
+    #### Request body
+    >>> {
+    ...     'id': int,
+    ...     'avatar': str,
+    ...     'name': str,
+    ...     'discriminator': int,
+    ...     'roles': List[int],
+    ...     'in_guild': bool
+    ... }
+
+    #### Status codes
+    - 200: returned on success
+    - 400: if the request body was invalid, see response body for details
+    - 404: if the user with the given `snowflake` could not be found
+
+    ### DELETE /bot/users/<snowflake:int>
+    Deletes the user with the given `snowflake`.
+
+    #### Status codes
+    - 204: returned on success
+    - 404: if a user with the given `snowflake` does not exist
+    """
+
+    serializer_class = UserSerializer
+    queryset = User.objects.prefetch_related('roles')
